Improve readme for release 1.0.0

+++ test/cases/Dao.php
@@ -19,7 +19,7 @@ final class DAOTest extends TestCase
                 id INTEGER PRIMARY KEY,
                 columnone VARCHAR(255),
                 columntwo INTEGER,
-                columnthree DOUBLE,
+                columnthree DOUBLE PRECISION,
                 active INTEGER DEFAULT 1
             )
         ");
